Core: Refactor bittorrent specific methods into common/bittorrent
