+++ lib/rucio/transfertool/bittorrent.py
@@ -17,10 +17,10 @@ import logging
 from os import path
 from typing import TYPE_CHECKING, Any, Optional
 
+from rucio.common.bittorrent import construct_torrent
 from rucio.common.config import config_get
 from rucio.common.constants import RseAttr
 from rucio.common.extra import import_extras
-from rucio.common.utils import construct_torrent
 from rucio.core.did_meta_plugins import get_metadata
 from rucio.transfertool.transfertool import TransferStatusReport, Transfertool, TransferToolBuilder
